LG-11302: Add aria labels to forms on hybrid handoff page
An accessibility report revealed that the two forms on this page have a
label of `""`. This is not great since they're both the same and empty.

This commit adds an aria label to them, which will make it more
accessible by adding the forms to the landmarks of the page.

Notes:

- I'm not sure the text for the button here is the best description of
the form. But, it already has translations and gives us a starting
point. We decided it was MVP.
- SimpleForm auto-created an id for one of the forms that I had to
override in order to make them consistent. See [this Slack thread](https://gsa-tts.slack.com/archives/C0NGESUN5/p1698872821571519) for
more details.
- The `hybrid_handoff` view is currently undergoing an A/B test. We
decided rather than create a new context for these tests or only include
them in either A or B, we would just duplicate the tests for now.

Co-authored-by: John Maxwell <john.maxwell@gsa.gov>
Co-authored-by: Brittany Greaner <brittany.greaner@gsa.gov>