Start writing decrypted PII to the session based on profile ID (#9515)
This commit builds on #9509. That commit added the ability to write PII to the session under the profile ID associated with that PII.

This commit starts using the functionality added in that PR to start writing both the pending and active profile to the session. Additionally in places where PII is written to the session outside the context of password authentication the corresponding profile ID is passed to the PII cacher.

[skip changelog]