Retire the `SsnForm`
The `SsnForm` class is a form object that does a lot of footwork to check if the SSN is unique and add that to the analytics. This form actually uses a service object to do that.

This commit removes the middleman and calls the service object in the controller (actually it is a concern shared between 2 controllers) that previuosly consumed this form.

[skip changelog]