LG-11537 Access the active profile in the userinfo endpoint
In #9509 we added the ability to specify which profile to fetch PII from when reading PII from the session.

This commit applies that capability to the OpenID Connect UserInfo endpoint to ensure that only the active profile's PII is used there.

The user info endpoint is used by service providers to access PII after auth. As a result it is not a request made by users and does not have a user session. To read PII we use the out-of-band session accessor's `#load_pii` method. This commit adds a `profile_id` arg to that method and passes in the active profile's ID when reading PII.

changelog: Internal, Active-pending profile, The active profile is now read from the session on the userinfo endpoint.