LG-11538 Show the PII from the active profile on the completions screen (#9599)
In #9509 we added the ability to specify which profile to fetch PII from when reading PII from the session.

This commit uses the active profile's ID to fetch PII on the completions screen. The completions screen should always be showing PII of the active profile.

changelog: Internal, Active-Pending profile, The active profile's PII is displayed on the completions screen.