LG-11534 Load the active profile from the session on broken personal key
In #9509 we added the ability to specify which profile to fetch PII from when reading PII from the session.

This commit uses the active profiles PII when encrypting recovery PII for the active profile when the active profile has a broken personal key.

changelog: Internal, Pending and active profile, The active profile PII is fetched with the PII cacher when a user with a broken personal key on their active profile signs in.