LG-11553 Remove recovery PII re-encryption from `PersonalKeyVerificationController`
The `PersonalKeyVerificationController` is used to verify a personal key as an MFA method and allow a user to sign in. When this is done a new personal key is issued.

This controller had code for re-encrypting the users profile with the newly issued personal key. However, a user with an active profile was never able to reach this path. The `check_personal_key_enabled` calls `TwoFactorAuthentication::PersonalKeyPolicy#enabled?`. This method returns false if the user has any profiles.

Since this code path is unreachable this commit removes it.

I was not able to find any tests covering this re-encryption behavior.

[skip changelog]