LG-11535 Encrypt the pending and active profile when a user updates their password (#9607)
In #9509 we added the ability to specify which profile to fetch PII from when reading PII from the session.

As a result of this change we can encrypt both the pending and active profile with user's password. This means both profiles can be decrypted on sign-in.

There are consequences for recovery with personal key. When the user changes their password and their data is encrypted their recovery PII is also encrypted as a consequence. The pending and active profile are both encrypted with a different personal key. In this commit I elected to display the pending profile personal key to the user so that profile is recoverable when it becomes active.

[skip changelog]