Add step_info for IPP controllers outside FSM.
StepInfo now takes the controller class as a parameter in the
initializer, then converts the controller to the full string name,
including modules, for use in url_for. This disambiguates between remote
idv and in person for ssn and verify_info controllers.

Removed controller_module from StepInfo since the above make it
unneeded.

Co-authored-by: Sonia Connolly <sonia.connolly@gsa.gov>
Co-authored-by: Alex Bradley <alexander.bradley@gsa.gov>