Modify the SP requirement IdV to require the SP to request verification (#9663)
In #4634 we restricted the identity verification process to require a service provider for a user to undergo proofing. This commit added other features like a `idv_sp_required` for configuring this behavior to be enabled or disabled. This commit required an SP to be present but did not require the SP to actually request verification.

This commit changes the code so that `IalContext` is invoked to check if verification was in fact requested. This way users can't go to an SP that does not require verification and start the verification process by navigating to `/verify`.

changelog: User-Facing Improvements, SP IdV Requirement, Users can only undergo proofing if the SP IAL context is for identity verification or greater.