LG-11688 Enforce the IdV available rule in a concern (#9665)
We have been discussing feature flags that can be toggled while the applicaiton is running. This would be an issue for the feature flags that drive `FeatureManagement#idv_available?`. That method is used in the routes file to handle routing for IdV routes when IdV is unavailable. This means that to change those feature flags while the app is running carries a requirement to redraw routes.

This commit makes a change to move equivalent logic to the routes file to the `Idv::Availability` concern. This concern is then included in all of the controller that were previously affected by the rule in the routes file.

changelog: Internal, IdV Availability, IdV availability is computed in a concern instead of the routes file