Remove background jobs and tooling for multi-region KMS migration
The multi-region KMS migration involved changes to migrate password digests and encrypted PII ciphertexts away from a KMS instance that used a single key to one that supports multi-region keys. This was done with a background job and tooling that would decrypt with the single region key and re-encrypt with the multi-region key.

All of the migrateable records have been migrated at this point so this job, its configuration, and the tooling to support it are not longer necessary.

[skip changelog]