Replace the `Pii::Cacher` implementation with the `Pii::ProfileCacher` implementation
The `Pii::ProfileCacher` was implemented in #9509. As described in that pull request:

> The Pii::ProfileCacher class implements the same API as Pii::Cacher and is intended to supersede Pii::Cacher when this work is done.

All of the invocations to `Pii::Cacher` that require the deprecated profile-unaware functionality have been removed and the new profile-aware functionality is fully implemented and in-user. This commit does the work of moving the `Pii::ProfileCacher` functionality into `Pii::Cacher` and removing `Pii::ProfileCacher`

This should not be merged until both `session_encrypted_profiles_read_enabled` and `session_encrypted_profiles_write_enabled` have been set to true. This commit removes those feature flags and assumes they are set to true.

changelog: Internal, Profile and session management, The pii cacher implementation was replaced with the profile cacher implementation