LG-11697 Store whether a biometric comparison is required in the SP session
We have not clearly defined how we want service providers to request a biometric comparison. This commit adds a query parameter to the OIDC authorization URL to specify that a biometric is required. This is temporary and intended to enable us to test the selfie / in-person requirement end-to-end before the API for requesting this functionality is fully defined.

This commit stores the attribute in the SP session. It does not make any changes to functionality based on the attribute or apply any changes to the SAML API.

The corresponding change on the sample app is here: https://github.com/18F/identity-oidc-sinatra/pull/149

changelog: Internal, Biometric comparison, A parameter was added to test the selfie workflow end-to-end