Workaround: fix headless mode used by Selenium locally
This fixes local dev for Chromedriver v120. We found that upgrading
Chromedriver to v120 in Docker for CI caused all sorts of issues, so for
the time being we are not upgrading CI.

This is a workaround from the failures in https://github.com/18F/identity-idp/pull/9734.

[See this documentation for details on the change](https://www.selenium.dev/blog/2023/headless-is-going-away/),
[this conversation in Slack for discussion](https://gsa-tts.slack.com/archives/C0NGESUN5/p1702068737905679),
and [this issue on Chromedriver](https://bugs.chromium.org/p/chromedriver/issues/detail?id=4440).