Add the `instant_verify_ab_test_discriminator` arg to the `ResolutionProofingJob`
This commit adds a arg to the `ResolutionProofingJob` which was used in the changes in #9743. That commit was reverted. Old workers did not know about this arg so they raised an ArgumentError when they encountered it. We reverted that change so we could introduce the arg in this one and then revert the revert.

[skip changelog]