Revert changes for LexisNexis InstantVerify AB test (#9764)
When we deployed the changes in #9763 and #9743 a new argument was introduced for the resolution proofing job. This argument was not recognized by old hosts performing the job which led to an argument error.

This commit reverts both of those and adds a commit to introduce the argument.

Once this commit is merged and deployed we will make another change to use the argument by reverting the reverts.

This commit adds a arg to the `ResolutionProofingJob` which was used in the changes in #9743. That commit was reverted. Old workers did not know about this arg so they raised an ArgumentError when they encountered it. We reverted that change so we could introduce the arg in this one and then revert the revert.

[skip changelog]