Add and test idv_selfie_image_file_uploaded
This event will only fire on selfie image upload, not on SDK
interaction.

For the metadata coming through on the event, I mostly modeled the data
I saw coming through for the front and back doc auth images, and then
removed some items that I didn't see data coming through for.

Unsure if that was the right way to go about it.

In addition, this is the only "selfie" event we can currently test on
the backend, so that is why it's the only one included in
`analytics_spec` for now. We may add a ticket for finding ways to test
the other events on the backend, especially since we are unsure if file
upload will stay as an option for selfie.