LG-11631: Add front end logging for selfie capture (#9795)
* changelog: Internal, Doc Auth, Add front end logging for selfie capture

This commit adds 5 logging events:

1. `idv_sdk_selfie_image_capture_opened` (**mobile only** - logs when a user opens the SDK for selfie capture)
2. `idv_sdk_selfie_image_capture_closed_without_photo` (**mobile only** - logs when a user starts to take a selfie with the SDK but exits out without adding one)
3. `idv_sdk_selfie_image_capture_failed` (**mobile only** - logs when an error happens in the process of adding the selfie via the SDK)
4. `idv_sdk_selfie_image_added` (**mobile only** - logs when a user uses the SDK to click the green checkmark and accept their selfie)
5. `idv_selfie_image_file_uploaded` (**desktop only** - logs when a user uses the file picker instead of the SDK to add a selfie)

We may eventually want to send more metadata of the selfie image as
well, but it is not currently available. So sending the metadata was
determined out of scope, and looked at in future work.

idv_selfie_image_file_uploaded is the only "selfie" event we can currently test on
the backend, so that is why it's the only one included in
`analytics_spec` for now. We may add a ticket for finding ways to test
the other events on the backend, especially since we are unsure if file
upload will stay as an option for selfie.

Co-authored-by: Charles Ferguson <charles.ferguson@gsa.gov>