Fix a 50/50 issue in the progress proofer (#9888)
A previous commit (https://github.com/18F/identity-idp/commit/e138a4904af3d7f4ed598d609a502fd8a8c1ae2a) changed the default values for `ipp_enrollment_in_progress` through stack for the vendor calls on the verify info step. This caused a number of `NoMethodError`s when the deploy happened. These errors results from the applicants address being nil when it was transformed for the LexisNexis API.

After some investigation it was discovered that the applicants address was being transformed for in-person proofing outside of in-person contexts. Since the user does not have the `identity_*` prefixed address values outside of in-person proofing this resulted in nil values for the address properties.

The cause of this is that `ipp_enrollment_in_progress` was introduced as a job argument with a default value of `true` which inverted it's previous value. As a result old web hosts were scheduling jobs without a `ipp_enrollment_in_progress` argument. New workers were picking up those jobs and erroneously assuming the default of `false.

This commit fixes all of the above.

[skip changelog]