LG-12015: stop spamming users who requested letters
If a user has an old GPO code that has received a reminder and a new
code that is not yet eligible for a reminder, we were sending them an
email every day. /o\ This was due to the join where a gpo code had a nil
reminder_sent_at and a gpo code created in the reminder window.

Now we only send an email if we find a gpo code that meets all
requirements.

changelog: Bug Fixes, Verify By Mail, No longer send email reminders to some users every day

Co-authored-by: Sonia Connolly <sonia.connolly@gsa.gov>