Stop merging things that aren't PII from the document into `pii_from_doc` (#9913)
* Stop merging things that aren't PII from the document into `pii_from_doc`

The `Idv::Session#pii_from_doc` method has a name that suggests it returns the PII that is read from the document. Prior to this commit this was not the case because we merge data that was not read from the document into the `pii_from_doc` hash. This was done as a convienence so it is available when communicating to vendors in downstream steps.

This commit removes the merge and provides the additional data explicitly in downsteam step.

[skip changelog]