LG-12143 Create a vector of trust parser
This commit introduces a service that can take a string representing a vector of trust and parse it into its components and values.

Vectors of trust are described in [RFC 8485](https://www.rfc-editor.org/rfc/rfc8485.html).

We are planning to use vectors of trust in our OIDC interface to allow service providers to describe the authentication and identity proofing feature set they need in place for their use case. This will be an alternative to ACR values. To that end this commit includes VoT 2 components with their own unique values.

The `P` component represents identity proofing. It contains the following values:

- `0`: No identity proofing
- `1`: Identity proofing without a biometric comparison (this is the current offering)
- `2`: Identity proofing with a biometric comparison (this is a future offering)

The `C` component represents credential usage. This essentially describes authentication features. It contains the following values:

- `0`: Default option. This maps to what occurs when AAL is set to "1".
- `1`: "No remember device". This maps to what occurs when AAL is set to "2".
- `2`: Phishing resistant auth methods
- `3`: PIV/CAC required

The construction of this mapping satisfies LG-12151.

As an example for how this new tool behaves, consider the vector "C2.P1". This would be parsed into `unphishable_mfa` for credential usage and `identity_proofing_no_biometric` for identity proofing.

[skip changelog]