Add the requested vectors of trust to `ServiceProviderRequest` (#9991)
* Add the requested vectors of trust to `ServiceProviderRequest`

We are working on implementing a feature for partners to request identity proofing and authentication features using vectors of trust. This will involve sending param describing the vector of trust in the original SAML or OIDC request. Within the context of OIDC this param is named `vtr`.

This commit adds a `vtr` property to `ServiceProviderRequest`. This property is unused and unset in the persisted service provider request. This will allow us to write to it in the future and initialize `ServiceProviderRequest`s with the value without resulting in an `ArgumentError` (thus avoiding a dreaded 50/50 state bug)

This commit also adds an `acr_values` property. This is looking forward to when parameters are consumed by the parser introduced in #9971.

[skip changelog]