LG-12143 Create a vector of trust parser (#9971)
* LG-12143 Create a vector of trust parser

This commit introduces a service that can take a string representing a vector of trust and parse it into its components and values.

Vectors of trust are described in RFC 8485.

We are planning to use vectors of trust in our OIDC interface to allow service providers to describe the authentication and identity proofing feature set they need in place for their use case. This will be an alternative to ACR values. To that end this commit includes VoT 2 components with their own unique values. Together these form 6 valid pairs of components and values.

The `P` component represents identity proofing. It contains the following values:

- `1`: Identity proofing is performed
- `2`: A biometric comparison is performed as part of identity proofing (this is a future offering)

The `C` component represents credential usage. This essentially describes authentication features. It contains the following values:

- `1`: Password authentication + MFA - Implied for all transactions
- `2`: AL2 conformant features engaged (e.g. no remember device)
- `a`: A phishing resistant authenticator is required
- `b`: A PIV or CAC is required

The construction of this mapping satisfies LG-12151.

As an example for how this new tool behaves, consider the vector "C2.P1". The "P1" vector implies the "C2" vector which is already included. The "C2" vector implies the "C1" vector. This results in the vector being expanded into "C1.C2.P1".

The result that is returned form the parser includes fields for the requirements for the authentication and identity proofing transaction. It includes the following requirements:

 - `aal2`: AAL2 is required
 - `phishing_resistant`: Phishing resistant authenticators are required
 - `hspd12`: PIV/CAC is required
 - `identity_proofing`: Identity proofing is required
 - `biometric_comparison`: A biometric comparison is required

This allows the result to be used to describe an authentication context that can be used to construct policies that ensure the requirements for the vector of trust are met.

[skip changelog]