LG-12190 Store vtr and acr_values in sp_session
This commit adds code to the `OpenidConnectAuthorizeForm` to consume a `vtr` param. This param validated and then added to the `ServiceProviderRequest` and eventually added to the `sp_session` by the `StoreSpMetadataInSession` service.

This `vtr` param will eventually be used along with the new `AuthnContextResolver` tooling to determine what features need to be in place for an authentication and identity proofing transaction.

[skip changelog]