Rename idv_phone_error_visited attrs
The controller did have a method that was `remaining_step`, but as it
was defined the exact same way as others named `remaining_attempt`, I
changed it to be consistent.

I tried creating a private method in the controller so that the method
could be used as the definition, instead of the same code 3x. But I
found doing so broke tests in an off-by-one sort of way - I think
somehow the private method held onto the value and was not updated the
same way as when it's defined each time.