Swap `#empty?` with `#blank?` and add tests
The commit that was un-reverted was problematic because of a `NoMethodError` when calling `empty?` on `nil` here:

```ruby
validates :acr_values, presence: true, if: ->(form) { form.vtr.empty? }
```

This commit replaces `#empty?` with `#blank?` which will work on nil. Additionally the `#parse_vtr` method was modified to consistently return `nil` if either a `vtr` param is not provided or the `vtr` param is not enabled.

Tests were added to cover this case.

[skip changelog]