Remove unused `@ial` ivar from `Users::SessionsController`
This ivar is part of a conditional that is used to determine if the link to authenticate with a PIV or CAC is visible.

It was added here: https://github.com/18F/identity-idp/commit/ccc40ed7e41e2f67063b3d7455b0990b0bd11bab#diff-cd2d90708a775a9fc605178934b17b6cb5d14d4396b06b916a84c43b1b2abd88R37

The computation for this value looks like this:

```ruby
def sp_session_ial
  sp_session[:ial].presence || 1
end
```

This will always have a truthy value unless the requested IAL is IALMax. After reading through the commit history I beleive this IALMax behavior is unintentional.

I believe this value was originally intended to require the user to enter a password if identity proofing was required. It appears this requirement changed but the ivar was left behind. This commit removes it.

[skip changelog]