Return a default result from `#resolved_authn_context_result` when no SP is present
Currently when there is no SP present `ApplicationController#resolved_authn_context_result` raises an error when it tries to resolve an Authn Context without a VoT or ACR values. This creates an issue when this method is called outside the context of a service provider.

In order to let the method be called outside of an SP context this method introduces a no-SP result which can be return when no SP is present. This allows the method to be called and requirements for authentication and identity-verification to be inspected without a service provider.

This method opts to base the conditional on whether an SP is present rather than whether ACR values or a VoT is present. This preserves behavior where a missing ACR value and VoT will cause an error to be raised. In the case where those values are missing and an SP is present we want to raise an error rather than proceeding with a no-SP result.

[skip changelog]