Recognize SAML IALMax using authn context comparison in SAML protocol
SAML SPs can request IALMax by using an IAL authn context that does not require proofing and including a Authn Context Comparison value of "minimum". This essentially says that the "IAL1" Authn context is the minimum acceptable context so "IAL2" is also acceptable.

This is problematic for the new `AuthnContextResolver` since it does not have visibility into these attributes on the SAML request when resolving the authn context.

This commit addresses the issue by returning the IALMax Authentication Context Reference from the `Saml#ial` in this case. This way it is picked up downstream by the `AuthnContextResolver`.

[skip changelog]