Recognize SAML IALMax using authn context comparison in SAML protocol (#10095)
SAML SPs can request IALMax by using an IAL authn context that does not require proofing and including a Authn Context Comparison value of "minimum". This essentially says that the "IAL1" Authn context is the minimum acceptable context so "IAL2" is also acceptable.

This is problematic for the new `AuthnContextResolver` since it does not have visibility into these attributes on the SAML request when resolving the authn context.

This commit addresses the issue by returning the IALMax Authentication Context Reference from `Saml#ial` in this case. This way the IALMax value is picked up downstream by the `AuthnContextResolver`.

This type of request has a few differences from an IALMax request with the IALMax authn context reference:

- If a service provider that cannot make IAL2 requests makes an IAL1 request in this way it will always be serviced without proofed attributes. An IALMax request with the IALMax authn context reference for a service provider that cannot make IAL2 requests results in an error.
- If a service provider is not on the list of SPs that can make an IALMax request but requests IALMax in this way it will be allowed. I am not sure if this is intended but it is the behavior prior to this commit so it is not a regression.

[skip changelog]