LG-12263 Remove legacy SP-sesison values from completions controller
The `sp_session[:ial2]` and `sp_session[:ialmax]` values are being retired in favor of `resolved_authn_context_result.identity_proofing?` and `resolved_authn_context_result.ialmax?`. This commit does the work of removing these 2 values from `SignUp::CompletionsController`. This removes the last of the references for `sp_session[:ialmax]` so once this is merged and deployed we can stop writing that value.

[skip changelog]