Fix a 50/50 issue in the IALMax code (#10141)
* Fix a 50/50 issue in the IALMax code

We introduced a change (ref: https://github.com/18F/identity-idp/pull/10095/files) that started storing the IALMAX ACR value when the IAL1 ACR value is sent with the minimum authn context comparison. Prior to this change the ACR value in the session would be IAL1.

That commit and later commits introduced code that expected that ACR value to be IALMAX when an IALMAX request was made. That led to issues when an IALMAX request was made with the old code that stored the IAL1 value. Specifically issues occurred in the SAML IDP controller when prompting to user to unlock their PII.

This commit makes the code aware of the old approach as well as the new one and handling both.

[skip changelog]