Use `sp_from_sp_session` for `ApplicationController#resolved_authn_context_result`
The `Application#resolved_authn_context_result` method uses the SP session to compute the resolved authn context. The `current_sp` method returns the current SP either by using the issuer in the SP sesison or by using the SP associated with a request ID param.

When `current_sp` is getting an SP associated with a request ID param then there is not an SP session. This leads to an error when computing the resolved authn context.

This change only passes the SP from the session into the `AuthnContextResolver` when computing the resolved authn context.

[skip changelog]