LG-12265: Stop reading from sp_session[:piv_cac_requested]
In a previous commit the `resolved_authn_context_result` was introduced to return a `Vot::Parser::Result` object that described the requirements for the current SP request considering SP default options. This is intended to be used to replace the keys in the `sp_session` that serve this purpose including the `piv_cac_requested` key.

This commit replaces places where the `sp_session[:piv_cac_requested]` value is read with new reads to the `resolved_authn_context_result`.

[skip changelog]