Setup for LG-12307: Doc Side Acuant to TS
[skip changelog]

As someone who is fairly new to TS, I found it difficult to understand
as I was trying to pass values between a JSX and a TSX file.

I decided to timebox converting this file to TS to see if that would
make my other work easier. It was fairly straightforward (I think).