LG-11625 Redirect to new personal key screen instead of show personal key flash (#10122)
Prior to this commit there were 2 cases where a personal key was showed to a user in a flash message:

- When the user resets their password and recovers their PII with a personal key
- When the user changes their password from account screen

In both cases the user had to have an identity proofed account for the flash to appear.

The flash had some issues. It was a distinct experience from the typical personal key issuance. The flash was hard to miss and did not include the features we build to ensure users save their personal key.

This commit removes the flash message and redirects the user to the personal key screen when they need to be presented with a personal key in all cases.

changelog: User-Facing Improvements, Personal key, The personal key issuance was improved.