Modify the ID token builder to recognize JSON encoded VTR params (#10168)
The `identities.vtr` column has one of the following values:

1. A JSON-encoded representation of the `vtr` param if the `vtr` param is consumed and used
2. Nil

The `IdTokenBuilder` uses this value to add a `vot` to the ID token. Prior to this commit this code expected a string representation of a single VoT and not a JSON encoded list. This commit addresses the issue and adjustst the tests to have the correct expectations.

[skip changelog]