Add `SamlIdp::Request#requested_vtr_authn_context`
We are working on adding support to making requests with vectors of trust. In order to do so we need to recognize when SAML requests are made with a vector of trust.

This commit adds a `SamlIdp::Request#requested_vtr_authn_context` method. It is a similar implementation to `#requested_ial_authn_context` and `#requested_aal_authn_context`. This implementation differs by recognizing AuthnContexts that include a VTR and returns the contents of the first AuthnContext that includes a VTR.