Remove shared example from SamlIdpController tests
The `SamlIdpController` tests had a shared example for what happens when a users signs in with an already verified account. This shared example is not actually shared between any examples. It is only used for the IAL2 case. This commit goes ahead and breaks it out into individual tests instead of using a shared example.

[skip changelog]