Render an error when a biometric is requested with a vector of trust
In #9837 we modified the IdP to render a 406 if biometric comparison was requested but not allowed for the current environment.

This change applied to the query parameter method for requesting biometric comparison which was the only way to request biometric comparison at the time. Since then we have enabled biometric comparison requests using vectors of trust but did not port over the feature for blocking biometric comparison requests in environments where they are not allowed.

This commit applies the feature to requests that use a vector of trust to request biometric comparison.

[skip changelog]