Remove `Vot::Parser::Result#aal_level_requested?`
This method was used to compute the requested AAL level to support cases that had not started using the `#aal2`, `#hspd12?`, and `#phishing_resistant?` properties on `Vot::Parser::Result`.

The following PRs remove the final spots where the numeric AAL value was used:

- #10186
- #10187
- #10194

With those changes there are not more uses of this method and it can be safely removed.

[skip changelog]