Modidy `ServiceProviderSession#mfa_expiration_interval` to not require an argument (#10186)
We modified the `mfa_expiration_interval` to use the `AuthnContextResolver` result to determine the SP session expiration. It uses this to determine if a non-AAL2 SP has made an AAL2 request in order to use the AAL2 expiration.

Prior to this commit the result was passed in from the application controller. We do not need to do this, however, since the SP session and the SP are both already passed in to the decorator. This commit uses those to compute the resolved authn context in the decorator.


[skip changelog]