Fix issue where DEFAULT_AAL ACR value led to SP default AAL resolution
The `AuthnContextResolver` consumes ACR values and VTR values and determine what requirements are for identity and authentication. It also includes SP defaults in this resolution if a value is not requested.

The `AuthnContextResolver` was looking at whether the AAL ACR values contained the substring `aal` to deterimine if an AAL ACR value was requested. If none was requested it used the SP default AAL. `urn:gov:gsa:ac:classes:sp:PasswordProtectedTransport:duo` is the ACR value for the default AAL (i.e. AAL1) and it does not include the `aal` substring. As a result it was not recognized as a requested AAL value and SP defaults were applied.

This commit fixes the issue above by checking for requested AAL values that include the `aal` substring or are the default AAL value.

[skip changelog]